pyproject.toml
[tool.poetry]
name = "teardrops"
version = "0.3"
description = "A sad discord-bot"
authors = ["Vy <vyom.j@protonmail.com>"]
license = "BSD-3-Clause"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.1.0"
motor = "^3.1.1"
aiohttp = "^3.8.3"
wikipedia = "^1.4.0"
dnspython = "^2.3.0"
jishaku = "^2.5.1"
python-dotenv = "^0.21.1"
loguru = "^0.6.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pylint-runner = "^0.6.0"
black = "^22.12.0"

[tool.poe.tasks]
start = "python bot/main.py"
lint = "pylint_runner"
flake8 = "python -m flake8 --exclude .venv --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics --format='::error file=%(path)s,line=%(row)d,col=%(col)d::[flake8] %(code)s: %(text)s' --show-source"
black = "black ."
black-check = "black . --check"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
